Move measure directives to common
