+++ app/components/common/measureCommitteeOpposing/measureCommitteeOpposing.html
