+++ app/components/common/measureCommitteeOpposing/measureCommitteeOpposingDirective.js
